+++ src/registry/endpoint.js
@@ -47,14 +47,6 @@ class Endpoint {
 	update() {
 
 	}
-
-	failure() {
-
-	}
-
-	success() {
-
-	}
 }
 
 module.exports = Endpoint;