+++ test/integration/registry.spec.js
@@ -1,11 +1,35 @@
 const Promise = require("bluebird");
+const E = require("../../src/errors");
 const H = require("./helpers");
+const { protectReject } = require("../unit/utils");
 
 let userService = {
 	name: "users",
 	actions: {
-		find() {},
-		get() {}
+		find() {
+			return "Found";
+		},
+		get: {
+			visibility: "public",
+			handler() {
+				return "Got";
+			},
+		},
+		update: {
+			visibility: "protected",
+			handler() {
+				return "Updated";
+			},
+		},
+		remove: {
+			visibility: "private",
+			handler() {
+				return "Removed";
+			}
+		},
+		removeWrap(ctx) {
+			return this.actions.remove();
+		}
 	},
 	events: {
 		"user.created"() {},
@@ -60,12 +84,12 @@ const mailService = {
 	}
 };
 
-const master = H.createNode("master", []);
-const node1 = H.createNode("node-1", [userService]);
-
-let node2 = H.createNode("node-2", [userService, paymentService]);
-
 describe("Test service registry", () => {
+	const master = H.createNode("first", "master", []);
+	const node1 = H.createNode("first", "node-1", [userService]);
+
+	let node2 = H.createNode("first", "node-2", [userService, paymentService]);
+
 	beforeAll(() => master.start());
 	afterAll(() => master.stop());
 
@@ -82,6 +106,12 @@ describe("Test service registry", () => {
 			expect(H.hasAction(master, "users.find")).toBe(true);
 			expect(H.hasAction(master, "users.get")).toBe(true);
 
+			// Check visibility
+			expect(H.hasAction(master, "users.update")).toBe(false);
+			expect(H.hasAction(node1, "users.update")).toBe(true);
+			expect(H.hasAction(node1, "users.remove")).toBe(false);
+			expect(node1.getLocalService("users").actions.remove).toBeInstanceOf(Function);
+
 			expect(H.getActionNodes(master, "users.find")).toEqual(["node-1"]);
 			expect(H.getActionNodes(master, "users.get")).toEqual(["node-1"]);
 
@@ -90,6 +120,7 @@ describe("Test service registry", () => {
 			expect(H.getEventNodes(master, "user.created")).toEqual(["node-1"]);
 		});
 	});
+
 
 	it("start node2 with userService & payment service", () => {
 		return node2.start().delay(100).then(() => {
@@ -134,7 +165,7 @@ describe("Test service registry", () => {
 	});
 
 	it("node2 recreate with posts, paymentMod", () => {
-		node2 = H.createNode("node-2", [paymentModService, postService]);
+		node2 = H.createNode("first", "node-2", [paymentModService, postService]);
 
 		return node2.start().delay(100).then(() => {
 			let infoNode2 = H.getNode(master, "node-2");
@@ -173,4 +204,35 @@ describe("Test service registry", () => {
 			expect(H.getActionNodes(master, "mail.send")).toEqual([]);
 		});
 	});
+});
+
+describe("Test action visibilities", () => {
+	const master = H.createNode("second", "master", []);
+	const node1 = H.createNode("second", "node-1", [userService]);
+
+	beforeAll(() => Promise.all([master.start(), node1.start()]));
+	afterAll(() => Promise.all([master.stop(), node1.stop()]));
+
+	it("should call remotely", () => {
+		return Promise.all([
+			master.call("users.find").catch(protectReject).then(res => expect(res).toBe("Found")),
+			master.call("users.get").catch(protectReject).then(res => expect(res).toBe("Got")),
+			master.call("users.update").then(protectReject).catch(err => expect(err).toBeInstanceOf(E.ServiceNotFoundError)),
+			master.call("users.remove").then(protectReject).catch(err => expect(err).toBeInstanceOf(E.ServiceNotFoundError)),
+		]);
+	});
+
+	it("should call locally", () => {
+		return Promise.all([
+			node1.call("users.find").catch(protectReject).then(res => expect(res).toBe("Found")),
+			node1.call("users.get").catch(protectReject).then(res => expect(res).toBe("Got")),
+			node1.call("users.update").catch(protectReject).then(res => expect(res).toBe("Updated")),
+			node1.call("users.remove").then(protectReject).catch(err => expect(err).toBeInstanceOf(E.ServiceNotFoundError)),
+		]);
+	});
+
+	it("should call directly inside action", () => {
+		return node1.call("users.removeWrap").catch(protectReject).then(res => expect(res).toBe("Removed"));
+	});
+
 });